🩹 ES-24868 (filters) issue with similarly named nodes
When multiple leaf with the same name exists the "open" action may not work as expected because comparison is done on the leaf's value instead of the it's path value. To compare node by path, a post-process to compute "path" must be done. This is now fixed in the sinequa/atomic-angular (v0.0.118) library